refactor(openclaw-plugin): extract polling/daily-digest domain
